Хранить бэкенд рендера в объекте и порядок модов массивом индексов
Разбор компромиссов раунда 3: вместо кэшей, которые ускоряли узкий путь
ценой памяти или других операций, — варианты, уменьшающие саму работу.

Рендер: неизменяемый RendererHeader первым членом каждого struct Renderer
заменяет глобальный реестр на 8 слотов под мьютексом и lock-free кэш.
RendererGetBackend быстрее на 17-82% (9/9 пар), кадровые стадии в шуме,
статики меньше примерно на 280 B, в объекте +4 B. Уходят лимит в восемь
рендеров, гонка чтения неатомарного указателя кэша и неверный ответ для
живого незарегистрированного объекта. Тестовые заглушки стриминга
заменены нулевыми: чужой указатель теперь разыменовывается, нули дают AUTO.

Моды: один массив uint8_t loadedOrder вместо 64-битного loadSequence в
каждом слоте. GetLoaded быстрее в 32 раза, LaiueModHost меньше на 976 B.
Код регистрации сервисов не менялся; её замер зависит от выравнивания
функции в DLL (на main тот же патч дал -10%, на 9f1e16c +24%).

Тесты: девять одновременно живых рендеров, порядок модов после выгрузок
из середины. Стенды r3_43 и r3_52 для парных A/B-замеров. Release и Debug
CTest 57/57, границы архитектуры OK.

+++ benchmarks/streaming_benchmark.c
@@ -35,8 +35,9 @@ static uint32_t benchSampleCount;
 static volatile uint64_t benchSink;
 
 // Фиктивный рендерер: мир пуст, меши не создаются, ни один Renderer*-вызов
-// не исполняется. Указатель только хранится стримингом.
-static uint8_t benchRendererPlaceholder;
+// не исполняется. Указатель только хранится стримингом; нули читаются
+// диспетчером рендера как AUTO.
+static uint64_t benchRendererPlaceholder;
 
 static void WriteText(const char* text)
 {
